Updated ts-codegen to the latest version, generated TypeScript types
